requirements.txt
pytest
pyahocorasick